refactor([issue-1154]): move createBibleStore to services + load curated SNP markers from JSON
Two lib-hygiene moves from the /do:better audit:

1. createBibleStore (the per-work CRUD + merge storage factory doing file I/O
   under data/writers-room/) moves from lib/storyBible.js to
   services/bibleStore.js, so the lib stays a pure catalog of
   sanitizers/transformers. storyBible.js re-exports it (and now exports the
   three pure helpers the factory needs — SANITIZERS, sortKey, nowIso) so the
   characters/places/objects call sites are untouched.

2. lib/curatedGenomeMarkers.js drops from 2210 to ~170 lines: the ~116-entry
   SNP dataset moves to a co-located curatedGenomeMarkers.json loaded at module
   init; MARKER_CATEGORIES + the classification functions stay in the .js. The
   JSON was generated by serializing the live export, verified byte-identical
   to the original array.

New curatedGenomeMarkers.test.js (14 cases) covers the JSON-load shape
invariants + the previously-untested classifyGenotype/formatGenotype/
resolveApoeHaplotype logic. Behavior-preserving.

Refs #1154
